requirements.txt
anyio==4.6.2.post1
astroid==3.3.5
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
dill==0.3.9
gevent==24.10.3
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iqoptionapi @ https://github.com/iqoptionapi/iqoptionapi/archive/refs/heads/master.zip#sha256=25326cba8e21bc3bb082ab2bd21282f435ce25589d89499a92e1879559c4ca42
isort==5.13.2
joblib==1.4.2
mccabe==0.7.0
numpy==2.1.2
pandas==2.2.3
platformdirs==4.3.6
pycparser==2.22
pylint==3.3.1
python-dateutil==2.9.0.post0
python-telegram-bot==21.6
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
TA-Lib @ file:///C:/Users/SCENTECH%20SERVICES/Desktop/codes/iq_options/TA_Lib-0.4.29-cp311-cp311-win_amd64.whl#sha256=aab5075b1c6f2482c80e4561bb44b888815ca41affe519ad559ff1db9dc010f2
threadpoolctl==3.5.0
tomlkit==0.13.2
tzdata==2024.2
urllib3==2.2.3
websocket==0.2.1
websocket-client==0.56.0
zope.event==5.0
zope.interface==7.1.1
